requirements.txt
Red-DiscordBot
black 
isort
aiohttp
yt_dlp
pycln